python-ocr-service/requirements.txt
# Python OCR Service Dependencies

flask==3.0.0
flask-cors==4.0.0
easyocr==1.7.1
pillow==10.1.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
torch==2.1.0
torchvision==0.16.0
